requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
pydantic
python-dotenv
redis
rapidfuzz
jellyfish
jinja2
pinecone-client
tiktoken
httpx
pydantic-settings
langchain
langchain-openai
gunicorn